feat: Add GetRoles endpoint to commonHandler
This commit adds a new endpoint, GetRoles, to the commonHandler. The GetRoles endpoint retrieves a list of roles from the database and returns them as a JSON response. This functionality allows clients to fetch the available roles in the system.

The GetRoles endpoint is implemented in the commonHandler file, and it is registered under the /roles route in the Register function.

This commit also includes changes to the ports.go file in the common module. The CommonRepository and CommonService interfaces now include the GetRoles method, which is implemented in the commonRepository and commonService files respectively.

The GetRoles method in the commonRepository queries the roles table in the database and returns a list of roles as entity.CommonResponse objects. The GetRoles method in the commonService simply calls the GetRoles method in the repository.

These changes enable the application to retrieve and serve the roles data to clients.